backend/requirements.txt
asgiref==3.9.0
boto3==1.39.3
botocore==1.39.3
dj-database-url==3.0.1
Django==4.2.7
django-cors-headers==4.7.0
djangorestframework==3.16.0
jmespath==1.0.1
pillow==11.3.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-decouple==3.8
s3transfer==0.13.0
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
whitenoise==6.9.0
gunicorn==21.2.0